# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexo",
    "description": "API do Nexo. Este documento é o contrato: o cliente TypeScript do front é gerado a partir dele, nunca escrito à mão.",
    "version": "v1"
  },
  "paths": {
...
    "/lancamentos/{id}/cobrar": {
      "post": {
        "tags": [
          "Lançamentos"
        ],
        "summary": "Emite boleto e Pix para um lançamento",
        "description": "Cria a cobrança no PSP e guarda o link onde o cliente escolhe entre boleto e Pix. Recusa cobrar duas vezes o mesmo lançamento: a segunda emissão não substituiria a primeira, e o cliente receberia dois boletos.",
        "operationId": "CobrarLancamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoCobrado"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaComProblemas"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LancamentoCobrado": {
        "required": [
          "cobrancaId",
          "cobrancaUrl",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cobrancaId": {
            "type": "string"
          },
          "cobrancaUrl": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Problema": {
        "required": [
          "campo",
          "descricao",
          "sugestao",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "campo": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "sugestao": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "RespostaComProblemas": {
        "required": [
          "problemas"
        ],
        "type": "object",
        "properties": {
          "problemas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Problema"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}